update nidigital examples
