+++ examples/nidmm_measurement/measurement.py
@@ -4,7 +4,7 @@ import contextlib
 import logging
 import math
 import pathlib
-from typing import Any, Dict, Tuple
+from typing import Tuple
 
 import click
 import grpc
@@ -12,12 +12,15 @@ import nidmm
 from _helpers import (
     ServiceOptions,
     configure_logging,
+    create_session_management_client,
+    get_grpc_device_channel,
     get_service_options,
     grpc_device_options,
     str_to_enum,
-    verbosity_option,
     use_simulation_option,
+    verbosity_option,
 )
+from _nidmm_helpers import _create_nidmm_session
 
 import ni_measurementlink_service as nims
 
@@ -79,12 +82,7 @@ def measure(
         resolution_digits,
     )
 
-    session_management_client = nims.session_management.Client(
-        grpc_channel=measurement_service.get_channel(
-            provided_interface=nims.session_management.GRPC_SERVICE_INTERFACE_NAME,
-            service_class=nims.session_management.GRPC_SERVICE_CLASS,
-        )
-    )
+    session_management_client = create_session_management_client(measurement_service)
 
     with contextlib.ExitStack() as stack:
         reservation = stack.enter_context(
@@ -106,7 +104,8 @@ def measure(
             )
 
         session_info = reservation.session_info[0]
-        session = stack.enter_context(_create_nidmm_session(session_info))
+        grpc_device_channel = get_grpc_device_channel(measurement_service, nidmm)
+        session = stack.enter_context(_create_nidmm_session(session_info, grpc_device_channel))
         session.configure_measurement_digits(
             str_to_enum(FUNCTION_TO_ENUM, measurement_type),
             range,
@@ -123,36 +122,6 @@ def measure(
         absolute_resolution,
     )
     return (measured_value, signal_out_of_range, absolute_resolution)
-
-
-def _create_nidmm_session(
-    session_info: nims.session_management.SessionInformation,
-) -> nidmm.Session:
-    options: Dict[str, Any] = {}
-    if service_options.use_simulation:
-        options["simulate"] = True
-        options["driver_setup"] = {"Model": "4081"}
-
-    session_kwargs: Dict[str, Any] = {}
-    if service_options.use_grpc_device:
-        session_grpc_address = service_options.grpc_device_address
-
-        if not session_grpc_address:
-            session_grpc_channel = measurement_service.get_channel(
-                provided_interface=nidmm.GRPC_SERVICE_INTERFACE_NAME,
-                service_class="ni.measurementlink.v1.grpcdeviceserver",
-            )
-        else:
-            session_grpc_channel = measurement_service.channel_pool.get_channel(
-                target=session_grpc_address
-            )
-        session_kwargs["grpc_options"] = nidmm.GrpcSessionOptions(
-            session_grpc_channel,
-            session_name=session_info.session_name,
-            initialization_behavior=nidmm.SessionInitializationBehavior.AUTO,
-        )
-
-    return nidmm.Session(session_info.resource_name, options=options, **session_kwargs)
 
 
 @click.command
